Enhance security through implementation of refresh tokens
Due to the stateless nature of JWTs, user engagement is limited to the short expiration window of the JWT. This means users will have to sign back in very often and this leads to a poor user experience. Refresh tokens are used to reduce this effect. They are assigned to the user at the same time as the JWTs but take longer to expire and are HTTPOnly. If a user's JWT is about to expire, the refresh token is used to create a new JWT automatically for the user, extending the amount of time until another sign is required.