Enhance api and exception handling structure
The previous system used to send requests to external apis was brittle and inflexible. This was caused by a high degree of coupling between calling classes and the implementation of the web client services. Notably, the coupling caused a need to modify the web client services every time a new class was introduced.

To fix this, a single universal class was created (WebClientService) to be used by all controllers for external api requests. This was further augmented by the use of generics and separation of public vs non-public api requests.

During initial attempts at fixing the api problem, it was found that the exception handling system had a related problem. The exception handling was closely tied to the controllers, leading to more duplicated and complex code. To fix this, controller advice was used with a global exception handler to keep most of the exception handling within one class.

The controller testing classes were also changed due to less of a need for integration tests (handled better by cbot testing harness) and more of a need for unit tests. Work was started on converting the controller classes by implementing springs mock mvc functions and a mock web server replicating the called upon external web servers.