Optimize refresh jwt mapping and associated classes
The refresh jwt mapping now uses the principal instead of requiring the equivalent RefreshRequest class to pass the user details. The mapping has also been integrated with the GlobalExceptionHandler and has two related exceptions to throw: RefreshExpiredException and RefreshTokenNotFoundException. Additionally, the testing class is in the process of converting from an integration test to a unit test to maximize specificity of testing.