Modify refresh token system so tokens are stored with user in user repo
This removes the need for a second repo of refresh tokens that are tied to the user ids of associated users. This also works well with using principal as a controller parameter to access users and therefore access refresh tokens.

Additionally, the integration tests for RefreshController and RefreshService were converted to unit tests. Associated classes were modified to fit the changes to the refresh system api.