Add components to handle duplicate user sign-up attempts
Sing-up requests that had usernames which were already linked to existing users were not being handled and were causing problems when retrieving users of the same username. To fix this, a class was created to handle requests with duplicate data, a group of exceptions was created to mark these requests, and these components were integrated into the sign-up controller pathways.